[REFACTOR] #207 : 유저 엔터티 속성 수정 및 폴더 구조 변경
