+++ src/main/java/com/umc/cardify/config/exception/ErrorResponseStatus.java
@@ -35,7 +35,10 @@ public enum ErrorResponseStatus {
 	NOT_FOUND_IMAGE(4013, "업로드할 이미지가 존재하지 않습니다."),
 
 	// 5000 : Server connection 오류
-	SERVER_ERROR(5000, "서버와의 연결에 실패하였습니다.");
+	SERVER_ERROR(5000, "서버와의 연결에 실패하였습니다."),
+
+	// 6000: Spring Security 오류
+	TOKEN_NOT_FOUND(6000, "토큰을 얻을 수 없습니다.");
 
 	private final int code;
 	private final String message;
