+++ src/main/java/com/umc/cardify/domain/enums/AuthProvider.java
@@ -1,0 +1,5 @@
+package com.umc.cardify.domain.enums;
+
+public enum AuthProvider {
+    GOOGLE, KAKAO
+}
