--- src/main/java/com/umc/cardify/handler/CustomAuthenticationFailureHandler.java
@@ -1,4 +1,0 @@
-package com.umc.cardify.handler;
-
-public class CustomAuthenticationFailureHandler {
-}
