+++ src/main/resources/application.yml
@@ -2,13 +2,10 @@ server:
   port: 8080
 spring:
   datasource:
-    url: jdbc:mysql://${DB_HOST}:${DB_PORT}/${DB_NAME}?serverTimezone=Asia/Seoul&useLegacyDatetimeCode=false
+    url: jdbc:mysql://localhost:3306/cardify_local_db?serverTimezone=Asia/Seoul&useLegacyDatetimeCode=false
     driverClassName: com.mysql.cj.jdbc.Driver
-    username: ${DB_USER}
-    password: ${DB_PASSWORD}
-  data:
-    mongodb:
-      uri: mongodb+srv://cardify:${DB_PASSWORD}@cardify.pckle.mongodb.net/cardify?retryWrites=true&w=majority&appName=Cardify
+    username: cardify_local
+    password: cardify123!
   jpa:
     hibernate:
       ddl-auto: none
